+++ ttp_templates/utils/models.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel, StrictInt, StrictStr, StrictBool
+from pydantic import BaseModel, ConfigDict, StrictInt, StrictStr, StrictBool
 from typing import Union, List
 
 
@@ -35,8 +35,76 @@ class LldpNeighborRecord(BaseModel):
     remote_device_management_ip: Union[None, StrictStr]
 
 
+class BgpNeighborRecord(BaseModel):
+    model_config = ConfigDict(extra="allow")
+
+    name: StrictStr
+    description: Union[None, StrictStr]
+    state: Union[None, StrictStr]
+    local_address: Union[None, StrictStr]
+    local_interface: Union[None, StrictStr]
+    remote_address: Union[None, StrictStr]
+    local_as: StrictInt
+    remote_as: StrictInt
+    peer_group: Union[None, StrictStr]
+    import_policies: List[StrictStr]
+    export_policies: List[StrictStr]
+    prefix_list_in: Union[None, StrictStr]
+    prefix_list_out: Union[None, StrictStr]
+    router_id: Union[None, StrictStr]
+    peering_type: Union[None, StrictStr]
+    vrf: Union[None, StrictStr]
+    hold_time: Union[None, StrictInt]
+    keepalive: Union[None, StrictInt]
+    uptime_seconds: Union[None, StrictInt]
+    max_ttl: Union[None, StrictInt]
+    afi: List[StrictStr]
+    ipv4_unicast_prefixes_sent: Union[None, StrictInt] = None
+    ipv4_unicast_prefixes_received: Union[None, StrictInt] = None
+    ipv6_unicast_prefixes_sent: Union[None, StrictInt] = None
+    ipv6_unicast_prefixes_received: Union[None, StrictInt] = None
+    ipv4_multicast_prefixes_sent: Union[None, StrictInt] = None
+    ipv4_multicast_prefixes_received: Union[None, StrictInt] = None
+    ipv6_multicast_prefixes_sent: Union[None, StrictInt] = None
+    ipv6_multicast_prefixes_received: Union[None, StrictInt] = None
+    ipv4_mpls_vpn_prefixes_sent: Union[None, StrictInt] = None
+    ipv4_mpls_vpn_prefixes_received: Union[None, StrictInt] = None
+    ipv6_mpls_vpn_prefixes_sent: Union[None, StrictInt] = None
+    ipv6_mpls_vpn_prefixes_received: Union[None, StrictInt] = None
+    ipv4_labeled_unicast_prefixes_sent: Union[None, StrictInt] = None
+    ipv4_labeled_unicast_prefixes_received: Union[None, StrictInt] = None
+    ipv6_labeled_unicast_prefixes_sent: Union[None, StrictInt] = None
+    ipv6_labeled_unicast_prefixes_received: Union[None, StrictInt] = None
+    l2vpn_evpn_prefixes_sent: Union[None, StrictInt] = None
+    l2vpn_evpn_prefixes_received: Union[None, StrictInt] = None
+    l2vpn_vpls_prefixes_sent: Union[None, StrictInt] = None
+    l2vpn_vpls_prefixes_received: Union[None, StrictInt] = None
+    ipv4_mdt_prefixes_sent: Union[None, StrictInt] = None
+    ipv4_mdt_prefixes_received: Union[None, StrictInt] = None
+    ipv4_flow_spec_prefixes_sent: Union[None, StrictInt] = None
+    ipv4_flow_spec_prefixes_received: Union[None, StrictInt] = None
+    ipv6_flow_spec_prefixes_sent: Union[None, StrictInt] = None
+    ipv6_flow_spec_prefixes_received: Union[None, StrictInt] = None
+    ipv4_sr_te_prefixes_sent: Union[None, StrictInt] = None
+    ipv4_sr_te_prefixes_received: Union[None, StrictInt] = None
+    ipv6_sr_te_prefixes_sent: Union[None, StrictInt] = None
+    ipv6_sr_te_prefixes_received: Union[None, StrictInt] = None
+    link_state_prefixes_sent: Union[None, StrictInt] = None
+    link_state_prefixes_received: Union[None, StrictInt] = None
+
+
 class InventoryRecord(BaseModel):
     description: StrictStr
     module: StrictStr
     serial: StrictStr
     slot: StrictStr
+
+
+class ArpRecord(BaseModel):
+    ip: StrictStr
+    age: Union[StrictInt, StrictStr]
+    mac: StrictStr
+    interface: Union[None, StrictStr] = None
+    protocol: Union[None, StrictStr] = None
+    state: Union[None, StrictStr] = None
+    type: Union[None, StrictStr] = None